requirements.txt
web3==6.20.2
eth-account==0.11.1
eth-abi==4.0.0
parsimonious==0.9.0
pandas==1.5.3
numpy==1.23.5
scipy==1.9.1
requests==2.28.1
python-dotenv>=0.19.0
matplotlib==3.10.1